+++ lib/src/ui/calendar_navigator_header_section.dart
@@ -2,8 +2,8 @@ import 'package:clean_calendar/src/models/calendar_properties.dart';
 import 'package:clean_calendar/src/state/page_controller.dart';
 import 'package:flutter/material.dart';
 
-class CalendarControlSection extends StatelessWidget {
-  const CalendarControlSection(
+class CalendarNavigatorHeaderSection extends StatelessWidget {
+  const CalendarNavigatorHeaderSection(
       {Key? key,
       required this.calendarProperties,
       required this.pageControllerState})
@@ -32,6 +32,33 @@ class CalendarControlSection extends StatelessWidget {
                     .copyWith(day: 1)) !=
                 DateUtils.dateOnly(pageViewDateTime.copyWith(day: 1));
 
+        Icon navigatorResetButtonIcon = calendarProperties.headerProperties
+                .navigatorDecoration?.navigatorResetButtonIcon ??
+            const Icon(
+              Icons.calendar_today,
+            );
+
+        Icon navigateLeftButtonIcon = calendarProperties
+                .headerProperties.navigatorDecoration?.navigateLeftButtonIcon ??
+            const Icon(
+              Icons.arrow_back_ios,
+              size: 12,
+            );
+
+        Icon navigateRightButtonIcon = calendarProperties.headerProperties
+                .navigatorDecoration?.navigateRightButtonIcon ??
+            const Icon(
+              Icons.arrow_forward_ios,
+              size: 12,
+            );
+
+        Color? monthYearTextColor = calendarProperties
+            .headerProperties.monthYearDecoration?.monthYearTextColor;
+
+        TextStyle? monthYearTextStyle = calendarProperties
+                .headerProperties.monthYearDecoration?.monthYearTextStyle ??
+            Theme.of(context).textTheme.titleSmall;
+
         return Container(
           margin: const EdgeInsets.only(left: 24, right: 8),
           child: Row(
@@ -41,7 +68,9 @@ class CalendarControlSection extends StatelessWidget {
                 child: Text(
                   "${monthsSymbolsList[pageViewDateTime.month - 1]} ${pageViewDateTime.year}",
                   overflow: TextOverflow.ellipsis,
-                  style: Theme.of(context).textTheme.titleSmall,
+                  style: monthYearTextColor != null
+                      ? monthYearTextStyle?.copyWith(color: monthYearTextColor)
+                      : monthYearTextStyle,
                 ),
               ),
               Flexible(
@@ -58,10 +87,7 @@ class CalendarControlSection extends StatelessWidget {
                                         curve: Curves.ease);
                               },
                               padding: EdgeInsets.zero,
-                              // iconSize: 12,
-                              icon: const Icon(
-                                Icons.calendar_today,
-                              ),
+                              icon: navigatorResetButtonIcon,
                             ),
                           )
                         : const Expanded(child: SizedBox()),
@@ -72,10 +98,7 @@ class CalendarControlSection extends StatelessWidget {
                               duration: kTabScrollDuration, curve: Curves.ease);
                         },
                         padding: EdgeInsets.zero,
-                        iconSize: 12,
-                        icon: const Icon(
-                          Icons.arrow_back_ios,
-                        ),
+                        icon: navigateLeftButtonIcon,
                       ),
                     ),
                     Expanded(
@@ -85,10 +108,7 @@ class CalendarControlSection extends StatelessWidget {
                               duration: kTabScrollDuration, curve: Curves.ease);
                         },
                         padding: EdgeInsets.zero,
-                        iconSize: 12,
-                        icon: const Icon(
-                          Icons.arrow_forward_ios,
-                        ),
+                        icon: navigateRightButtonIcon,
                       ),
                     ),
                   ],